pyproject.toml
[project]
name = "manuals"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.13"
dependencies = []

[dependency-groups]
dev = [
    "bs4>=0.0.2",
    "dataclasses-json>=0.6.7",
    "html5lib>=1.1",
    "pip>=25.2",
    "pyyaml>=6.0.2",
    "show-in-file-manager>=1.1.5",
]

[tool.pyright]
include = ["scripts"]
extraPaths = ["archipelago"]
